requirements.txt
PyAudio==0.2.14
pynput==1.8.1
pyobjc-core==11.0
pyobjc-framework-ApplicationServices==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-CoreText==11.0
pyobjc-framework-Quartz==11.0
six==1.17.0
